Default vLLM dtype to bfloat16
Qwen3 sweeps already set vllm_dtype=bfloat16 explicitly; flip the default
so other runs don't silently stay on float16.

+++ vllm_colocate.py
@@ -27,7 +27,7 @@ class VLLMColocateClient:
     update_weights_from_model, set_scales, shutdown).
     """
 
-    def __init__(self, model_name, gpu_memory_utilization=0.05, dtype="float16"):
+    def __init__(self, model_name, gpu_memory_utilization=0.05, dtype="bfloat16"):
         # Disable vLLM subprocess so we can access the model directly
         os.environ["VLLM_ENABLE_V1_MULTIPROCESSING"] = "0"
         os.environ.setdefault("VLLM_LOGGING_LEVEL", "ERROR")
